Run asdf install in setup script
As a consumer of this template, after you run the Middleman command
to initialize a new project, we run the `bin/setup`. However, if you
don't have your Ruby version set to the version we expect, then the
script fails.

Checking for asdf (which is commonly used at thoughtbot) and then
running its `install` command, will that our required Ruby version
is in place while executing the setup script.