Compile Sass via external pipeline using Sass CLI
Ruby Sass is deprecated and the sassc gem is unable to find Bourbon.
The turnover in Sass compilers in the Ruby community continues to
cause friction.

This switches over to pulling Sass (Dart Sass, specifically) from npm
and runs its CLI directly. This has the added benefit of using the most
modern Sass compiler.

https://sass-lang.com/documentation/cli/dart-sass